// client/src/data/geoNested.ts
export interface GeoBBox {
	latMin: number;
	latMax: number;
	lonMin: number;
	lonMax: number;
	airports?: string[];
}

export interface GeoDistrict extends GeoBBox {
	code: string;
	nameZh: string;
}

export interface GeoCity extends GeoBBox {
	code: string;
	nameZh: string;
	districts?: GeoDistrict[];
}

export type NestedGeoMap = Record<string, Record<string, GeoCity[]>>;
// ...
export const NESTED_GEO_US: NestedGeoMap = {
  "CA": [
      { code: "CA-LA", nameZh: "洛杉矶", latMin: 33.7, latMax: 34.35, lonMin: -118.65, lonMax: -117.9, airports: ["LAX"] },
      { code: "CA-SF", nameZh: "旧金山湾区", latMin: 37.45, latMax: 37.85, lonMin: -122.55, lonMax: -121.9, airports: ["SFO", "SJC", "OAK"] },
      { code: "CA-SD", nameZh: "圣迭戈", latMin: 32.6, latMax: 33.1, lonMin: -117.3, lonMax: -116.9, airports: ["SAN"] },
    ],
  "NY": [
      { code: "NY-NYC", nameZh: "纽约都会", latMin: 40.55, latMax: 40.9, lonMin: -74.25, lonMax: -73.7, airports: ["JFK", "LGA", "EWR"], districts: [{ code: "NY-MAN", nameZh: "曼哈顿", latMin: 40.7, latMax: 40.82, lonMin: -74.02, lonMax: -73.9 }, { code: "NY-QUE", nameZh: "皇后区", latMin: 40.65, latMax: 40.78, lonMin: -73.95, lonMax: -73.75 }, { code: "NY-BRK", nameZh: "布鲁克林", latMin: 40.58, latMax: 40.7, lonMin: -74.05, lonMax: -73.9 }] },
    ],
  "TX": [
      { code: "TX-DFW", nameZh: "达拉斯沃斯堡", latMin: 32.6, latMax: 33.1, lonMin: -97.2, lonMax: -96.5, airports: ["DFW", "DAL"] },
      { code: "TX-HOU", nameZh: "休斯顿", latMin: 29.6, latMax: 30.05, lonMin: -95.65, lonMax: -95.1, airports: ["IAH", "HOU"] },
    ],
  "FL": [
      { code: "FL-MIA", nameZh: "迈阿密", latMin: 25.7, latMax: 26.0, lonMin: -80.35, lonMax: -80.1, airports: ["MIA", "FLL"] },
      { code: "FL-ORL", nameZh: "奥兰多", latMin: 28.35, latMax: 28.6, lonMin: -81.5, lonMax: -81.2, airports: ["MCO"] },
    ]
};

export const NESTED_GEO_JP: NestedGeoMap = {
  "JP-KNT": [
      { code: "TYO-AREA", nameZh: "东京", latMin: 35.55, latMax: 35.85, lonMin: 139.55, lonMax: 139.9, airports: ["HND", "NRT"], districts: [{ code: "TYO-C23", nameZh: "都心", latMin: 35.65, latMax: 35.72, lonMin: 139.68, lonMax: 139.78 }, { code: "TYO-C24", nameZh: "湾岸", latMin: 35.6, latMax: 35.68, lonMin: 139.75, lonMax: 139.9 }] },
    ],
  "JP-KNS": [
      { code: "OSA-CITY", nameZh: "大阪", latMin: 34.6, latMax: 34.75, lonMin: 135.4, lonMax: 135.6, airports: ["KIX", "ITM"] },
    ]
};

export const NESTED_GEO_DE: NestedGeoMap = {
  "DE-BY": [
      { code: "DE-MUC", nameZh: "慕尼黑", latMin: 48.1, latMax: 48.25, lonMin: 11.45, lonMax: 11.7, airports: ["MUC"] },
    ],
  "DE-BE": [
      { code: "DE-BER", nameZh: "柏林", latMin: 52.45, latMax: 52.58, lonMin: 13.3, lonMax: 13.55, airports: ["BER"] },
    ]
};